+++ pkg/scad/booleans/difference.go
