+++ pkg/scad/values/bool.go
