+++ pkg/scad/values/floatxyz.go
