+++ pkg/scad/values/string.go
